requirements.txt
# Core dependencies
PyYAML>=6.0
pandas>=1.3.0
matplotlib>=3.5.0

# Development dependencies
pytest>=6.0.0
pytest-cov>=2.0.0
black>=21.0.0
flake8>=3.8.0
mypy>=0.800
